Refactors the repository code into a module
Because:
- Clean up
