+++ lib/devlogs.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
+require_relative "devlogs/gem"
 require_relative "devlogs/cli"
