+++ lib/devlogs/gem.rb
@@ -1,0 +1,5 @@
+module Devlogs
+  def self.lib_root
+    File.join(File.dirname(__dir__), "devlogs")
+  end
+end
